refactor: paths
