+++ config.yml
@@ -1,6 +1,9 @@
 title:                  "Bootstrap"
 baseURL:                "https://getbootstrap.com"
 
+# TODO(HiDeoo) Fix path
+docsDir:                "site-new"
+
 # TODO(HiDeoo) Check if still relevant, update Astro config and prolly remove this value from here
 publishDir:             "_site"
