+++ site-new/src/components/Scripts.astro
@@ -1,6 +1,7 @@
 ---
 import type { HTMLAttributes } from 'astro/types'
-import { getConfig, getVersionedBsJsPath } from '../libs/config'
+import { getConfig } from '../libs/config'
+import { getVersionedBsJsPath } from '../libs/path'
 import type { Layout } from '../libs/layout'
 import StackBlitz from '../components/StackBlitz.astro'
