+++ site-new/src/components/footer/Footer.astro
@@ -1,6 +1,7 @@
 ---
 import BootstrapWhiteFillIcon from '../icons/BootstrapWhiteFillIcon.astro'
-import { getConfig, getVersionedDocsPath } from '../../libs/config'
+import { getConfig } from '../../libs/config'
+import { getVersionedDocsPath } from '../../libs/path'
 ---
 
 <footer class="bd-footer py-4 py-md-5 mt-5 bg-body-tertiary">
