+++ site-new/src/components/head/Favicons.astro
@@ -1,5 +1,5 @@
 ---
-import { getVersionedDocsPath } from '../../libs/config'
+import { getVersionedDocsPath } from '../../libs/path'
 ---
 
 <link rel="apple-touch-icon" href={getVersionedDocsPath('/assets/img/favicons/apple-touch-icon.png')} sizes="180x180" />
