+++ site-new/src/components/head/Social.astro
@@ -1,5 +1,6 @@
 ---
-import { getConfig, getVersionedDocsPath } from '../../libs/config'
+import { getConfig } from '../../libs/config'
+import { getVersionedDocsPath } from '../../libs/path'
 import { getStaticImageSize } from '../../libs/image'
 import type { Layout } from '../../libs/layout'
