+++ site-new/src/components/head/Stylesheet.astro
@@ -1,6 +1,7 @@
 ---
 import type { HTMLAttributes } from 'astro/types'
-import { getConfig, getVersionedBsCssPath } from '../../libs/config'
+import { getConfig } from '../../libs/config'
+import { getVersionedBsCssPath } from '../../libs/path'
 import type { Layout } from '../../libs/layout'
 import Scss from './Scss.astro'
