+++ site-new/src/components/header/Navigation.astro
@@ -1,6 +1,7 @@
 ---
 import type { CollectionEntry } from 'astro:content'
-import { getConfig, getVersionedDocsPath } from '../../libs/config'
+import { getConfig } from '../../libs/config'
+import { getVersionedDocsPath } from '../../libs/path'
 import type { Layout } from '../../libs/layout'
 import BootstrapWhiteFillIcon from '../icons/BootstrapWhiteFillIcon.astro'
 import GitHubIcon from '../icons/GitHubIcon.astro'
