+++ site-new/src/components/home/CSSVariables.astro
@@ -1,5 +1,5 @@
 ---
-import { getVersionedDocsPath } from '../../libs/config'
+import { getVersionedDocsPath } from '../../libs/path'
 import Code from '../shortcodes/Code.astro'
 ---
