+++ site-new/src/components/home/GetStarted.astro
@@ -1,5 +1,6 @@
 ---
-import { getConfig, getVersionedDocsPath } from '../../libs/config'
+import { getConfig } from '../../libs/config'
+import { getVersionedDocsPath } from '../../libs/path'
 import Code from '../shortcodes/Code.astro'
 ---
