+++ site-new/src/components/home/MastHead.astro
@@ -1,5 +1,6 @@
 ---
-import { getConfig, getVersionedDocsPath } from '../../libs/config'
+import { getConfig } from '../../libs/config'
+import { getVersionedDocsPath } from '../../libs/path'
 import Ads from '../Ads.astro'
 import Code from '../shortcodes/Code.astro'
 ---
