+++ site-new/src/components/home/Plugins.astro
@@ -1,5 +1,5 @@
 ---
-import { getVersionedDocsPath } from '../../libs/config'
+import { getVersionedDocsPath } from '../../libs/path'
 import { getData } from '../../libs/data'
 import Code from '../shortcodes/Code.astro'
