+++ site-new/src/components/home/Themes.astro
@@ -1,5 +1,6 @@
 ---
-import { getConfig, getVersionedDocsPath } from '../../libs/config'
+import { getConfig } from '../../libs/config'
+import { getVersionedDocsPath } from '../../libs/path'
 import DropletFillIcon from '../icons/DropletFillIcon.astro'
 ---
