+++ site-new/src/layouts/SingleLayout.astro
@@ -1,9 +1,9 @@
 ---
-import Ads from '../components/Ads.astro';
+import Ads from '../components/Ads.astro'
 import BaseLayout from './BaseLayout.astro'
 
 interface Props {
-  description: string,
+  description: string
   title: string
 }
