+++ site-new/src/libs/config.ts
@@ -29,6 +29,7 @@ const configSchema = z.object({
   current_ruby_version: zVersionSemver,
   description: z.string(),
   docs_version: zVersionMajorMinor,
+  docsDir: z.string(),
   download: z.object({
     dist: z.string().url(),
     dist_examples: z.string().url(),
@@ -75,40 +76,6 @@ export function getConfig(): Config {
 
     throw new Error('Failed to load configuration from `config.yml`', { cause: error })
   }
-}
-
-export function getVersionedDocsPath(path: string): string {
-  const { docs_version } = getConfig()
-
-  return `/docs/${docs_version}/${path.replace(/^\//, '')}`
-}
-
-export function getVersionedBsCssPath(direction: 'rtl' | undefined): string {
-  let bsCssLinkHref = '/dist/css/bootstrap'
-
-  if (direction === 'rtl') {
-    bsCssLinkHref = `${bsCssLinkHref}.rtl`
-  }
-
-  if (import.meta.env.PROD) {
-    bsCssLinkHref = `${bsCssLinkHref}.min`
-  }
-
-  bsCssLinkHref = `${bsCssLinkHref}.css`
-
-  return getVersionedDocsPath(bsCssLinkHref)
-}
-
-export function getVersionedBsJsPath(): string {
-  let bsJsScriptSrc = '/dist/js/bootstrap.bundle'
-
-  if (import.meta.env.PROD) {
-    bsJsScriptSrc = `${bsJsScriptSrc}.min`
-  }
-
-  bsJsScriptSrc = `${bsJsScriptSrc}.js`
-
-  return getVersionedDocsPath(bsJsScriptSrc)
 }
 
 type Config = z.infer<typeof configSchema>
