+++ site-new/src/libs/image.ts
@@ -1,9 +1,9 @@
 import path from 'node:path'
 import sizeOf from 'image-size'
-import { getDocsStaticPath } from './astro'
+import { getDocsStaticFsPath } from './path'
 
 export function getStaticImageSize(imagePath: string) {
-  const size = sizeOf(path.join(getDocsStaticPath(), imagePath))
+  const size = sizeOf(path.join(getDocsStaticFsPath(), imagePath))
 
   if (!size.height || !size.width) {
     throw new Error(`Failed to get size of static image at '${imagePath}'.`)
