+++ site-new/src/pages/docs/[version]/index.astro
@@ -1,6 +1,7 @@
 ---
 import RedirectLayout from '../../../layouts/RedirectLayout.astro'
-import { getConfig, getVersionedDocsPath } from '../../../libs/config'
+import { getConfig } from '../../../libs/config'
+import { getVersionedDocsPath } from '../../../libs/path'
 
 export function getStaticPaths() {
   return [
