+++ site-new/src/pages/docs/index.astro
@@ -1,6 +1,6 @@
 ---
 import RedirectLayout from '../../layouts/RedirectLayout.astro'
-import { getVersionedDocsPath } from '../../libs/config'
+import { getVersionedDocsPath } from '../../libs/path'
 ---
 
 <RedirectLayout path={getVersionedDocsPath('/getting-started/introduction/')} />
